# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ChecklistHQ API
  description: >-
    JSON API for ChecklistHQ (auth, checklists, notifications, users,
    templates, groups, tenant admin). Generated by hand from
    internal/api/*.go — see DESIGN.md for the broader system design.


    Auth model: cookie-based sessions (`checklists_session`, HttpOnly) set by
    login/register/password-reset-confirm. Every non-GET/HEAD request also
    requires a `X-CSRF-Token` header matching the (non-HttpOnly)
    `checklists_csrf` cookie value — a double-submit CSRF check — EXCEPT
    `/login`, `/register`, `/password-reset/request`, and
    `/password-reset/confirm`, which are unauthenticated by design and exempt.
  version: "1.0"
servers:
  - url: http://localhost:8080
    description: Local API (default api_port)
security:
  - sessionCookie: []
...
paths:
...
  /api/checklists:
    post:
      tags: [Checklists]
      summary: Create a checklist
      description: >-
        Supply `template_id` to instantiate the checklist from that
        template's latest version.
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateChecklistRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Checklist" }
        "400": { description: Bad request (invalid assignment, validation error) }
        "403": { description: Checklist creation is restricted and caller isn't allowed to create one }
    get:
      tags: [Checklists]
      summary: List checklists relevant to the caller
      description: >-
        Returns checklists the caller created, approves, is directly
        assigned, or could claim (unclaimed group assignment).
      parameters:
        - name: status
          in: query
          required: false
          schema: { type: string, enum: [open, validating, complete] }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Checklist" }
        "400": { description: Invalid status value }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: checklists_session
      description: >-
        Set by /login, /register, and /password-reset/confirm. Every
        non-safe-method request also needs a matching X-CSRF-Token header —
        see the CsrfToken parameter on each mutating operation.
  parameters:
...
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: { type: string }
      description: Must equal the value of the checklists_csrf cookie (double-submit CSRF check).
  schemas:
...
    ChecklistItem:
      type: object
      properties:
        ID: { type: integer, format: int64 }
        ChecklistID: { type: integer, format: int64 }
        Name: { type: string }
        Position: { type: integer }
        Checked: { type: boolean }
        CheckedBy: { type: integer, format: int64, nullable: true }
        CheckedAt: { type: string, format: date-time, nullable: true }
        ValidationRef: { type: string }
        AssigneeOverrideUserID:
          type: integer
          format: int64
          nullable: true
          description: Set only via the reject flow — points at whoever originally checked this item.
    Checklist:
      type: object
      properties:
        ID: { type: integer, format: int64 }
        TenantID: { type: integer, format: int64 }
        TemplateID: { type: integer, format: int64 }
        Name: { type: string }
        CreatorID: { type: integer, format: int64 }
        AssignedGroupID: { type: integer, format: int64, nullable: true }
        AssignedUserID: { type: integer, format: int64, nullable: true }
        Hidden: { type: boolean }
        ApproverID: { type: integer, format: int64, nullable: true }
        Status: { type: string, enum: [open, validating, complete] }
        CreatedAt: { type: string, format: date-time }
        Items:
          type: array
          items: { $ref: "#/components/schemas/ChecklistItem" }
...
    CreateChecklistRequest:
      type: object
      properties:
        template_id:
          type: integer
          format: int64
          description: The checklist is instantiated from this template's latest version.
        name:
          type: string
          description: Optional; defaults to the template's own name if omitted/empty.
        assigned_group_id: { type: integer, format: int64, nullable: true }
        assigned_user_id: { type: integer, format: int64, nullable: true }
        hidden: { type: boolean }
        approver_id: { type: integer, format: int64, nullable: true }
      required: [template_id]
...